Feat/optimize server resource

+++ app/_lib/hooks/useTrack.ts
@@ -1,14 +1,14 @@
+// /app/_lib/hooks/useTrack.ts
 import { fetchWithToken } from '@/app/_lib/auth/fetchWithToken'
+import { useCurrentUser } from '@/app/_lib/hooks/useCurrentUser'
+import { deleteTrackFromIndexedDB, getTrackFromIndexedDB, setTrackToIndexedDB } from '@/app/_lib/indexedDB'
 import { state } from '@/app/_lib/state'
 import { DeleteTrackAPI, GetTracksAPI } from '@/app/types/api'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { useSnapshot } from 'valtio'
 
-/**
- * 트랙 관리 커스텀 훅
- * 인증 상태에 따라 서버 API 또는 로컬 상태 사용
- */
 export const useTrack = () => {
+    const { isMember } = useCurrentUser()
     const snapshot = useSnapshot(state)
 
     const queryClient = useQueryClient()
@@ -25,44 +25,49 @@ export const useTrack = () => {
     const createTrackMutation = useMutation({
         mutationFn: async (file: File) => {
             try {
-                // 1. 프리사인드 URL 요청
-                const presignedResponse = await fetchWithToken('/api/upload/presigned-url', {
-                    method: 'POST',
-                    body: JSON.stringify({
-                        fileName: file.name,
-                        fileType: file.type,
-                    }),
-                })
-
-                // 2. S3에 파일 업로드
-                const uploadResponse = await fetch(presignedResponse.url, {
-                    method: 'PUT',
-                    body: file,
-                    headers: {
-                        'Content-Type': file.type,
-                    },
-                })
-
-                if (!uploadResponse.ok) throw new Error('S3 upload failed')
-
-                // 3. 메타데이터 저장 요청
+                // 1. db 저장 요청
                 const playlistId = snapshot.UI.currentPlaylistId
-                const s3Url = presignedResponse.url.split('?')[0]
 
-                return await fetchWithToken('/api/track/create', {
+                const response = await fetchWithToken('/api/track/create', {
                     method: 'POST',
                     body: JSON.stringify({
                         fileName: file.name,
-                        url: s3Url,
                         playlistId,
                     }),
                 })
+
+                setTrackToIndexedDB(response.id, file)
+
+                if (isMember) {
+                    // 2. 프리사인드 URL 요청
+                    const presignedResponse = await fetchWithToken('/api/upload/presigned-url', {
+                        method: 'POST',
+                        body: JSON.stringify({
+                            id: response.id,
+                            fileName: response.fileName,
+                            fileType: file.type,
+                        }),
+                    })
+
+                    // 3. S3에 파일 업로드
+                    const uploadResponse = await fetch(presignedResponse.url, {
+                        method: 'PUT',
+                        body: file,
+                        headers: {
+                            'Content-Type': file.type,
+                        },
+                    })
+
+                    if (!uploadResponse.ok) throw new Error('S3 upload failed')
+                }
+
+                return response
             } catch (error) {
                 throw new Error(`File upload failed: ${error instanceof Error ? error.message : 'Unknown error'}`)
             }
         },
-        onSuccess: (data) => {
-            state.UI.focusedTrackId = data.id
+        onSuccess: (response) => {
+            state.UI.focusedTrackId = response.id
         },
         onSettled: () => {
             queryClient.invalidateQueries({ queryKey })
@@ -73,9 +78,13 @@ export const useTrack = () => {
     // 트랙 삭제 뮤테이션
     const deleteTrackMutation = useMutation({
         mutationFn: async (id: string) => {
-            return fetchWithToken(`/api/track/${id}/delete`, {
-                method: 'DELETE',
-            }) as Promise<DeleteTrackAPI['Response']>
+            deleteTrackFromIndexedDB(id)
+
+            if (isMember) {
+                return fetchWithToken(`/api/track/${id}/delete`, {
+                    method: 'DELETE',
+                }) as Promise<DeleteTrackAPI['Response']>
+            }
         },
         onMutate: async (id) => {
             await queryClient.cancelQueries({ queryKey })
@@ -95,17 +104,31 @@ export const useTrack = () => {
         },
     })
 
-    // 단일 트랙의 blob url | presigned URL 가져오기 함수
-    const getTrackUrl = async (id: string): Promise<string> => {
-        const data = await fetchWithToken(`/api/track/${id}/get`, {
-            method: 'GET',
-        })
+    // 단일 트랙의 blob | presigned URL => blob 가져오기 함수
+    const getTrackBlobUrl = async (id: string): Promise<Blob> => {
+        const blob = await getTrackFromIndexedDB(id)
+        if (blob) {
+            return blob
+        } else {
+            if (isMember) {
+                const response = await fetchWithToken(`/api/track/${id}/presigned-url`, {
+                    method: 'GET',
+                })
 
-        if (data.error) {
-            throw new Error('Failed to fetch track presigned URL')
+                if (response.error) {
+                    throw new Error('Failed to fetch track presigned URL')
+                }
+
+                const fileResponse = await fetch(response.presignedUrl)
+                const blob = await fileResponse.blob()
+
+                setTrackToIndexedDB(id, blob)
+
+                return blob
+            } else {
+                throw new Error('Track not found')
+            }
         }
-
-        return data.presignedUrl
     }
 
     return {
@@ -116,6 +139,6 @@ export const useTrack = () => {
         isCreating: createTrackMutation.isPending,
         deleteTrack: deleteTrackMutation.mutate,
         isDeleting: deleteTrackMutation.isPending,
-        getTrackUrl,
+        getTrackBlobUrl,
     }
 }
